refactor: remove onnxruntime dependence in OcrLite.h
