+++ include/OcrLite.h
@@ -1,14 +1,12 @@
 #ifndef __OCR_LITE_H__
 #define __OCR_LITE_H__
 
-#include "opencv2/core.hpp"
-#include <onnxruntime/core/session/onnxruntime_cxx_api.h>
+#include <string>
 #include "OcrStruct.h"
-#include "DbNet.h"
-#include "AngleNet.h"
-#include "CrnnNet.h"
 
-class OcrLite {
+class OcrLiteImpl;
+
+class OcrLite{
 public:
     OcrLite();
 
@@ -35,28 +33,8 @@ public:
                      int padding, int maxSideLen,
                      float boxScoreThresh, float boxThresh, float unClipRatio, bool doAngle, bool mostAngle);
 
-    OcrResult detectImageBytes(const uint8_t *data, long dataLength, int grey, int padding, int maxSideLen,
-                               float boxScoreThresh, float boxThresh, float unClipRatio, bool doAngle, bool mostAngle);
-    OcrResult detectBitmap(uint8_t *bitmapData, int width, int height,int channels, int padding, int maxSideLen,
-                           float boxScoreThresh, float boxThresh, float unClipRatio, bool doAngle, bool mostAngle);
-
 private:
-    bool isOutputConsole = false;
-    bool isOutputPartImg = false;
-    bool isOutputResultTxt = false;
-    bool isOutputResultImg = false;
-    FILE *resultTxt;
-    DbNet dbNet;
-    AngleNet angleNet;
-    CrnnNet crnnNet;
-    char *loggerBuffer;
-    std::vector<cv::Mat> getPartImages(cv::Mat &src, std::vector<TextBox> &textBoxes,
-                                       const char *path, const char *imgName);
-
-    OcrResult detect(const char *path, const char *imgName,
-                     cv::Mat &src, cv::Rect &originRect, ScaleParam &scale,
-                     float boxScoreThresh = 0.6f, float boxThresh = 0.3f,
-                     float unClipRatio = 2.0f, bool doAngle = true, bool mostAngle = true);
+    OcrLiteImpl* pImpl;
 };
 
 #endif //__OCR_LITE_H__
