+++ src/OcrLiteJni.cpp
